+++ CHANGELOG.md
@@ -5,6 +5,14 @@ Changelog
 
 Features:
 
+  - [config] JSON schema support
+  - [cli] `phpactor config:init` command to create or update config (to
+    include JSON schema location)
+
+## 2022-01-03 (0.18.0)
+
+Features:
+
   - [language-server] Import all names refactoring - @dantleech
   - [language-server] Extract expression - @BladeMF
   - [language-server] Extract method generation - @BladeMF